refactor(src): group modules into core/adapters/app/cli layers
Groups every source module by its relationship to the outside world so
"where does a new module go?" reads off the tree in one glance. The
entry point becomes main.ts, freeing the cli name for the delivery
folder, and the cli- filename prefix drops now that the folder carries
it. Import paths only — no module's internals change, and the version
module's package.json lookup is adjusted for its new depth.
