+++ src/classify.test.ts
@@ -4,8 +4,8 @@ import {
   lastResultLine,
   parseResultEvent,
   reclassifyCorrelatedFailures,
-} from "./classify.js";
-import type { WorkerOutcome } from "./types.js";
+} from "./core/classify.js";
+import type { WorkerOutcome } from "./core/types.js";
 
 describe("classifyInfrastructure", () => {
   it("classifies a usage-limit death", () => {
