+++ src/cli/docs.ts
