+++ src/config.test.ts
@@ -1,5 +1,5 @@
 import { describe, expect, it } from "vitest";
-import { ConfigError, modelForAttempt, resolveConfig } from "./config.js";
+import { ConfigError, modelForAttempt, resolveConfig } from "./core/config.js";
 
 describe("resolveConfig", () => {
   it("takes the scope parent and max_workers from the config file", () => {
