+++ src/core/plan.ts
